move node:https SNICallback tests to their own file
node-http.test.ts has a host-dependent ECONNREFUSED in 'request via http
proxy, issue#4295' (proxy listens on localhost, client dials 127.0.0.1);
keep the SNICallback coverage in a dedicated file alongside the other
node-https-* tests so it stays green independent of that.
